Gson serialization is almost perfect.
Deserialized numbers are created as the lowest sized primitive it can be
without loss of precission or data. This doesn't affect getting values
in a compound as all primitives are dynamically cast to the primitive
you're trying to get.

More GDF stuff, dynamic casting of int[] <-> byte[] in Compounds.

Added some test stuff as well.